a bunch of changes:
moved outputDir into CompilationOptions, because it seems to make sense there instead of being forced to be passed everywhere
added the ability to specify the output dir on the command line with -o or --outputDir
wrote compileTurtles which takes an input file and compiles all the way to assembly, writing intermediate files
added a ReplTools module that will be a place a put a bunch of stuff i use for testing in the repl.
added the ability for tests to search from more than one directory.