bunch of runtime computer related changes:
fixed up displaying the computer state on an error
fixed up push and pop to work with Pointers, because rsp is always a pointer to memory.
changed runOp to be an operation on Runtime objects instead of just ints.
fixed L1Interp so that the tests pass. this required delicate checking of which operations should use pointers vs nums, and calling runOp in other place.
did some code cleanups that the intellij plugin suggested